+++ chaosmagpy/data_utils.py
@@ -645,10 +645,10 @@ def gauss_units(deriv=None):
         'nT'
 
     >>> gauss_units(1)
-        '$\\\\mathrm{nT}\\\\cdot\\\\mathrm{yr}^{-1}$'
+        '$\\\\mathrm{nT}\\\\mathrm{yr}^{-1}$'
 
     >>> gauss_units(2)
-        '$\\\\mathrm{nT}\\\\cdot\\\\mathrm{yr}^{-2}$'
+        '$\\\\mathrm{nT}\\\\mathrm{yr}^{-2}$'
 
     """
 
@@ -657,6 +657,6 @@ def gauss_units(deriv=None):
     if deriv == 0:
         units = 'nT'
     else:
-        units = '$\\mathrm{{nT}}\\cdot\\mathrm{{yr}}^{{{:}}}$'.format(-deriv)
+        units = '$\\mathrm{{nT}}\\mathrm{{yr}}^{{{:}}}$'.format(-deriv)
 
     return units